+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `check_ratio_limits` function parameters updated
 - `check_temperature_limits` function parameters updated
 - `check_grind_limits` function parameters updated
+- `calc_coffee` function parameters updated
 - Test system modified
 - `Python 3.6` support dropped
 ## [1.7] - 2025-03-26